backend/requirements.txt
google-adk>=1.0.0
google-genai>=1.0.0
google-api-python-client>=2.100.0
websockets>=13.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
